fix(money): switch Product.price and Order.priceSum to BigDecimal
Float / MySQL FLOAT stores approximations, so accumulating order
totals silently drifts (0.1 + 0.2 != 0.3) and reconciliation fails.
Move the money columns to DECIMAL(10, 2) and the Java fields to
BigDecimal so every price stored, read, and summed is exact.

V2__price_columns_to_decimal.sql runs after V1 to ALTER MODIFY the
existing FLOAT columns to DECIMAL(10, 2) in one shot; the app is not
yet deployed, so no two-phase rollout is required. Update
OrderController to accumulate line totals via BigDecimal.multiply and
add, tighten ProductController's productPrice regex to accept up to
two fractional digits, and switch the DTOs (OrderItemDetail,
CustomOrderResponse) plus the tests to BigDecimal literals.

Document the reasoning and BigDecimal usage rules (String constructor,
compareTo not equals, divide with RoundingMode, MySQL DECIMAL sizing)
in docs/money-types.md.

+++ src/main/java/com/albertkingdom/shoppingwebsite/model/CustomOrderResponse.java
@@ -1,18 +1,19 @@
 package com.albertkingdom.shoppingwebsite.model;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 
 public class CustomOrderResponse {
     private Long id;
-    private Float priceSum;
+    private BigDecimal priceSum;
     private Long userId;
     private String userEmail;
     private List<OrderItemDetail> orderItemDetailList;
 
 
-    public CustomOrderResponse(Long id, Float priceSum, Long userId, String userEmail, List<OrderItemDetail> orderItemDetailList) {
+    public CustomOrderResponse(Long id, BigDecimal priceSum, Long userId, String userEmail, List<OrderItemDetail> orderItemDetailList) {
         this.id = id;
         this.priceSum = priceSum;
         this.userId = userId;
@@ -32,7 +33,7 @@ public class CustomOrderResponse {
         return priceSum;
     }
 
-    public void setPriceSum(Float priceSum) {
+    public void setPriceSum(BigDecimal priceSum) {
         this.priceSum = priceSum;
     }
